Refactor components and active items functionality
